refactor: extract stream mode handlers
