+++ src/js/config/mode-docs.js
@@ -1,0 +1,11 @@
+export const MODE_DOCS = {
+  boon: { file: 'docs/api/modes.md#boon' },
+  bane: { file: 'docs/api/modes.md#bane' },
+  bone: { file: 'docs/api/modes.md#bone' },
+  bonk: { file: 'docs/api/modes.md#bonk' },
+  honk: { file: 'docs/api/modes.md#honk' },
+  boof: { url: 'https://example.com/boof-api', file: 'docs/api/modes.md#boof' },
+  lore: { file: 'docs/api/modes.md#lore' },
+  focal: { file: 'docs/api/modes.md#focal' },
+  passive: { file: 'docs/api/modes.md#passive' },
+};
