+++ src/js/ui/modes/bane.js
@@ -1,0 +1,4 @@
+import { ConfigModeHandler } from './base.js';
+export class BaneModeHandler extends ConfigModeHandler {
+  constructor(container) { super(container, 'bane'); }
+}
