+++ src/js/ui/modes/bone.js
@@ -1,0 +1,4 @@
+import { ConfigModeHandler } from './base.js';
+export class BoneModeHandler extends ConfigModeHandler {
+  constructor(container) { super(container, 'bone'); }
+}
