+++ src/js/ui/modes/bonk.js
@@ -1,0 +1,4 @@
+import { ConfigModeHandler } from './base.js';
+export class BonkModeHandler extends ConfigModeHandler {
+  constructor(container) { super(container, 'bonk'); }
+}
