+++ src/js/ui/modes/honk.js
@@ -1,0 +1,4 @@
+import { ConfigModeHandler } from './base.js';
+export class HonkModeHandler extends ConfigModeHandler {
+  constructor(container) { super(container, 'honk'); }
+}
