+++ src/js/ui/modes/lore.js
@@ -1,0 +1,4 @@
+import { ConfigModeHandler } from './base.js';
+export class LoreModeHandler extends ConfigModeHandler {
+  constructor(container) { super(container, 'lore'); }
+}
